+++ mod_redstoneExtended/src/net/minecraft/src/redstoneExtended/Laser/TileEntityLaser.java
@@ -4,8 +4,8 @@ import net.minecraft.src.NBTTagCompound;
 import net.minecraft.src.TileEntity;
 
 public class TileEntityLaser extends TileEntity {
-    public byte orientation = 0;
-    public short distance = 0;
+    public int orientation = 0;
+    public int distance = 0;
 
     public net.minecraft.src.redstoneExtended.Laser.LaserMode mode = new net.minecraft.src.redstoneExtended.Laser.LaserMode();
 
@@ -23,8 +23,8 @@ public class TileEntityLaser extends TileEntity {
     @Override
     public void writeToNBT(NBTTagCompound nbtTagCompound) {
         super.writeToNBT(nbtTagCompound);
-        nbtTagCompound.setByte("Orientation", orientation);
-        nbtTagCompound.setShort("Distance", distance);
+        nbtTagCompound.setByte("Orientation", (byte)orientation);
+        nbtTagCompound.setShort("Distance", (short)distance);
         NBTTagCompound modeTag = new NBTTagCompound();
         net.minecraft.src.redstoneExtended.Laser.LaserMode.writeToNBT(modeTag, mode);
         nbtTagCompound.setCompoundTag("Mode", modeTag);
